django/auth_service/requirements.txt
Django
djangorestframework
django-allauth
python-dotenv
django-cors-headers
djangorestframework-simplejwt